+++ js/jquery.terminal-src.js
@@ -3881,7 +3881,7 @@
         // :: Function split and strips single and double quotes
         // :: and escapes spaces
         // ---------------------------------------------------------------------
-        split_arguments: function(string) {
+        split_arguments: function split_arguments(string) {
             return $.map(string.match(command_re) || [], function(arg) {
                 return $.terminal.parse_argument(arg, false);
             });
@@ -3896,10 +3896,59 @@
         // ---------------------------------------------------------------------
         // :: Same as parse_command but arguments are parsed using split_arguments
         // ---------------------------------------------------------------------
-        split_command: function(string) {
+        split_command: function split_command(string) {
             return process_command(string, function(arg) {
                 return $.terminal.parse_argument(arg, false);
             });
+        },
+        // ---------------------------------------------------------------------
+        // :; function that parse arguments like yargs library
+        // ---------------------------------------------------------------------
+        parse_options: function parse_options(arg, options) {
+            var settings = $.extend({}, {
+                boolean: []
+            }, options);
+            if (typeof arg === 'string') {
+                return parse_options($.terminal.split_arguments(arg), options);
+            }
+            var result = {
+                _: []
+            };
+            function token(value) {
+                this.value = value;
+            }
+            var rest = arg.reduce(function(acc, arg) {
+                if (arg.match(/^-/) && acc instanceof token) {
+                    result[acc.value] = true;
+                }
+                if (arg.match(/^--/)) {
+                    var name = arg.replace(/^--/, '');
+                    if (settings.boolean.indexOf(name) == -1) {
+                        return new token(name);
+                    } else {
+                        result[name] = true;
+                    }
+                } else if (arg.match(/^-/)) {
+                    var single = arg.replace(/^-/, '').split('');
+                    if (settings.boolean.indexOf(single.slice(-1)[0]) == -1) {
+                        var last = single.pop();
+                    }
+                    single.forEach(function(single) {
+                        result[single] = true;
+                    });
+                    if (last) {
+                        return new token(last);
+                    }
+                } else if (acc instanceof token) {
+                    result[acc.value] = arg;
+                } else if (arg) {
+                    result._.push(arg);
+                }
+            }, null);
+            if (rest instanceof token) {
+                result[rest.value] = true;
+            }
+            return result;
         },
         // ---------------------------------------------------------------------
         // :: function executed for each text inside [{ .... }]