+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Features
 * add ascii_table utility in separated file
 * per user command line history
+* add $.terminal.parse_options which return same object as yargs parser
 
 ### Bugfix
 * disable history in read & login (regression from 1.16.0 history interpreter option)